Muestro decimales para el monto en vez de aproximado

+++ MercadoEnvio/Clases/Estadistica.cs
@@ -3,6 +3,6 @@
     public class Estadistica
     {
         public string Username { get; set; }
-        public int Extra { get; set; }
+        public double Extra { get; set; }
     }
 }
